// src/api/models/blockResDTOKind.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type BlockResDTOKind = typeof BlockResDTOKind[keyof typeof BlockResDTOKind];


export const BlockResDTOKind = {
  GROUP_UNCATEGORIZED: 'GROUP_UNCATEGORIZED',
  GROUP: 'GROUP',
  EXPERIENCE: 'EXPERIENCE',
  SECTION_DETAIL: 'SECTION_DETAIL',
  SECTION_ACHIEVEMENT: 'SECTION_ACHIEVEMENT',
  SECTION_TASK: 'SECTION_TASK',
  SECTION_PROBLEM_SOLVING: 'SECTION_PROBLEM_SOLVING',
  SECTION_LEARNING: 'SECTION_LEARNING',
  CONTENT: 'CONTENT',
} as const;