// src/api/models/createBlockReqDTOKind.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type CreateBlockReqDTOKind = typeof CreateBlockReqDTOKind[keyof typeof CreateBlockReqDTOKind];


export const CreateBlockReqDTOKind = {
  GROUP_UNCATEGORIZED: 'GROUP_UNCATEGORIZED',
  GROUP: 'GROUP',
  EXPERIENCE: 'EXPERIENCE',
  SECTION_DETAIL: 'SECTION_DETAIL',
  SECTION_ACHIEVEMENT: 'SECTION_ACHIEVEMENT',
  SECTION_TASK: 'SECTION_TASK',
  SECTION_PROBLEM_SOLVING: 'SECTION_PROBLEM_SOLVING',
  SECTION_LEARNING: 'SECTION_LEARNING',
  CONTENT: 'CONTENT',
} as const;